requirements.txt
torch==2.6.0
streamlit==1.44.1
numpy==1.25.0
pandas==1.5.3
faiss-cpu==1.10.0  
transformers==4.51.3
sentence-transformers==4.1.0
Pillow==9.4.0      
requests==2.32.3   
openai==1.75.0
dotenv